V1.1.0.0
Added toggle button next to each part in the PartsList that toggles
whether that part is used in EC consumption/production calculations.
Added Dark-Side Calculations GUI. This allows you to select a Celestial
Body, Input a Target Orbit height and AmpYear will calculate the
dark-side period in Hours:Mins:Secs and also show you the amount of EC
you would require to run all vessel parts that consume EC for that
period (it Does NOT include any Power Production parts in this
calculation).